Move some cosntants to a separate file and fix some missing jsdoc types
