+++ src/language.js
@@ -1,66 +1,7 @@
+// @ts-check
 const fuzzysort = require('fuzzysort');
+const { languageMap, LanguagesEnum, languages, languageCodes, localeMap, languageDict } = require('./constants');
 
-const LanguagesEnum = {
-	'ChineseSimplified'  : 'ChineseSimplified',
-	'ChineseTraditional' : 'ChineseTraditional',
-	'English'            : 'English',
-	'French'             : 'French',
-	'German'             : 'German',
-	'Indonesian'         : 'Indonesian',
-	'Italian'            : 'Italian',
-	'Japanese'           : 'Japanese',
-	'Korean'             : 'Korean',
-	'Portuguese'         : 'Portuguese',
-	'Russian'            : 'Russian',
-	'Spanish'            : 'Spanish',
-	'Thai'               : 'Thai',
-	'Turkish'            : 'Turkish',
-	'Vietnamese'         : 'Vietnamese'
-};
-
-// converts Genshin's language codes into expanded strings
-const languageMap = {
-	'CHS': 'ChineseSimplified',
-	'CHT': 'ChineseTraditional',
-	'DE':  'German',
-	'EN':  'English',
-	'ES':  'Spanish',
-	'FR':  'French',
-	'ID':  'Indonesian',
-	'IT':  'Italian',
-	'JP':  'Japanese',
-	'KR':  'Korean',
-	'PT':  'Portuguese',
-	'RU':  'Russian',
-	'TH':  'Thai',
-	'TR':  'Turkish',
-	'VI':  'Vietnamese'
-};
-// array of language strings
-const languages = Object.values(languageMap);
-// array of language codes
-const languageCodes = Object.keys(languageMap);
-
-// // converts expanded strings into javascript locale codes
-const localeMap = {
-	'ChineseSimplified':  'zh-cn',
-	'ChineseTraditional': 'zh-tw',
-	'German':             'de',
-	'English':            'en',
-	'Spanish':            'es',
-	'French':             'fr',
-	'Indonesian':         'id',
-	'Italian':            'it',
-	'Japanese':           'ja',
-	'Korean':             'ko',
-	'Portuguese':         'pt',
-	'Russian':            'ru',
-	'Thai':               'th',
-	'Turkish':            'tr',
-	'Vietnamese':         'vi'
-};
-
-const languageDict = Object.keys(languageMap).concat(Object.values(languageMap));
 
 function autocompleteLanguage(input, dict) {
     let result = fuzzysort.go(input, dict, { limit: 1 })[0];
@@ -89,10 +30,10 @@ function format(langs) {
 }
 
 module.exports = {
-	LanguagesEnum: LanguagesEnum,
-	languages: languages, // array of strings
-	languageCodes: languageCodes,
-	languageMap: languageMap,
-	localeMap: localeMap,
-	format: format
+	LanguagesEnum,
+	languages, // array of strings
+	languageCodes,
+	languageMap,
+	localeMap,
+	format
 };
